crud_with_postgres/requirements.txt
python-dotenv
fastapi
uvicorn
sqlalchemy
pydantic